code cleanup
cleaned up various code
Decided that the rename of the package was best solution, since doing some changes to building.
Also since this hasn't had a public mod.io release this should be fine as this plugin has gone virtually unused.
It will break any plugins if used as a dependency(unlikely)
Version bump to 2.0.0 due to the breaking nature of package change