Setup for Doxygen Documentation generation
Rework of the comments to support Doxygen documentation, with the desired output
Added usage references, params and return value information
Added examples for events
Reworked the classes to match the LUA functions, which makes documentation nicer.  This functional has no impact.
Fixed ResetPlayers to be an Action rather than function with a return bool